# backend/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Help you to deliver better software",
        "title": "Code Scanner API",
        "termsOfService": "http://twitter.com/sutanto1010/",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
        "/crud/{table}/v1": {
            "get": {
                "description": "This is generic CRUD endpoint, can be use to do CRUD for existing table or future table",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Generic CRUD",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Table name",
                        "name": "table",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    }
                }
            }
        },
        "/request-to-scan/v1": {
            "post": {
                "description": "Use this endpoint to make scan request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Request to scan",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.Response": {
            "type": "object",
            "properties": {
                "Data": {},
                "Message": {
                    "type": "string"
                },
                "Success": {
                    "type": "boolean"
                },
                "Total": {
                    "type": "integer"
                }
            }
        }
    }
}